requirements.txt
flask
flask-sqlalchemy
python-dotenv
google-generativeai